Parse control commands incrementally.

+++ eventlog-socket/include/eventlog_socket.h
@@ -22,7 +22,7 @@ typedef void eventlog_socket_control_command_handler_t(
     const eventlog_socket_control_command_t command, void *user_data);
 
 bool eventlog_socket_control_register_namespace(
-    size_t namespace_len, const char namespace[namespace_len + 1],
+    uint8_t namespace_len, const char namespace[namespace_len + 1],
     eventlog_socket_control_namespace_id_t *namespace_id_out);
 
 bool eventlog_socket_control_register_command(
